% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{best_split}
\alias{best_split}
\title{Function to calculate best split}
\usage{
best_split(response, treated, x, agg_op, ate = 0, minbucket = 100)
}
\description{
Function to calculate best split
}
