% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumulative_te}
\alias{cumulative_te}
\title{calculate cumulative ATE for each potential split}
\usage{
cumulative_te(r, t, x, min_tc = 1L)
}
\arguments{
\item{r}{response vector}

\item{t}{indicator variable for treated}

\item{x}{numeric vector of covariated, sorted low to high}

\item{min_tc}{smallest number of treated / control units we need
before considering a split}
}
\value{
matrix of left / right daughter ATEs
}
\description{
calculate cumulative ATE for each potential split
}
