% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumulative_tscore}
\alias{cumulative_tscore}
\title{calculate cumulative ATE t-score for each potential split}
\usage{
cumulative_tscore(r, t, x, ate = 0, minbucket = 100L)
}
\arguments{
\item{r}{response vector}

\item{t}{indicator variable for treatment}

\item{x}{numeric predictor vector (sorted low to high)}

\item{ate}{average treatment effect to test}

\item{minbucket}{only consider splits with at least minbucket obs}
}
\value{
matrix of left / right daughter tscores
}
\description{
calculate cumulative ATE t-score for each potential split
}
