% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_functions.R
\name{partitions}
\alias{partitions}
\title{create all possible partitions}
\usage{
partitions(n)
}
\arguments{
\item{n}{number of elements in set}
}
\value{
list containing partitions
}
\description{
create all possible partitions
}
