% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{subgroup_perm_test}
\alias{subgroup_perm_test}
\title{permutation test for depth 1 trees}
\usage{
subgroup_perm_test(response, treated, X, direction = c("max", "min"),
  n_perm = 1000, ate = 0, minbucket = 100)
}
\arguments{
\item{response}{numeric outcome of interest}

\item{treated}{boolean vector of treatment assignments}

\item{X}{data.frame of predictors; must be either numeric or factor types}

\item{direction}{"max" to search for largest treatment effect, "min" to
search for smallest treatment effect}

\item{n_perm}{number of permutations to calculate}

\item{ate}{null average treatment effect to test against}

\item{minbucket}{minumum number of observations required to consider a split}
}
\description{
Perform a permutation test for a null-effect for the largest ATE discovered
from a search using a depth-1 subgroup tree.
}
