% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup_tree.R
\name{subgroup_tree}
\alias{subgroup_tree}
\title{subgroup detection tree}
\usage{
subgroup_tree(response, treated, X, direction = c("max", "min"), ...)
}
\arguments{
\item{response}{numeric outcome of interest}

\item{treated}{boolean vector of treatment assignments}

\item{X}{data.frame of predictors; must be either numeric or factor types}

\item{direction}{"max" to search for largest treatment effect, "min" to
search for smallest treatment effect}

\item{...}{additional arguments to rpart.control, such as maxdepth, etc.}
}
\description{
Fits a tree designed to aggressively seek out subsets of the data with
large (or small) average treatment effects
}
\examples{
\dontrun{

set.seed(123)
n = 500
p = 5
treated = sample(c(TRUE, FALSE), n, replace = TRUE)
X = as.data.frame(matrix(rnorm(n * p), n))
high_ate = X[, 1] > 0.5
response = rbinom(n, 1, ifelse(high_ate & treated, 0.9, 0.5))
max_tree = subgroup_tree(response, treated, X, 'max', maxdepth = 2,
                        minbucket = 50)
print(max_tree)
}
}
