% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{tscore}
\alias{tscore}
\title{compute a t-score for the ATE on a subset of the data defined by "cond"}
\usage{
tscore(response, treated, cond, ate = 0)
}
\description{
compute a t-score for the ATE on a subset of the data defined by "cond"
}
